+++ apps/demo/app/index.tsx
@@ -1,11 +1,8 @@
 import { useXUIColors, useXUITheme } from '@xaui/native/core'
-import { StyleSheet, View, Text } from 'react-native'
-import { Button } from '@xaui/native/button'
-import { Progress } from '@xaui/native/progress'
-import { ActivityIndicator } from '@xaui/native/indicator'
-import { Checkbox } from '@xaui/native/checkbox'
-import { Switch } from '@xaui/native/switch'
+import { StyleSheet, View } from 'react-native'
 import { useState, useEffect } from 'react'
+import { Select, SelectItem } from '@xaui/native/select'
+import { Button } from '@xaui/native/button'
 
 export default function HomeScreen() {
   const colors = useXUIColors()
@@ -30,74 +27,28 @@ export default function HomeScreen() {
         },
       ]}
     >
-      <Button
-        variant="solid"
-        size="xs"
-        spinnerPlacement="end"
-        themeColor="primary"
-        startContent={<Text>🚀</Text>}
-        fullWidth
-      >
-        Azure limpide
-      </Button>
-      <Button
-        variant="solid"
-        size="sm"
-        spinnerPlacement="end"
-        isLoading
-        themeColor="secondary"
-        startContent={<Text>🚀</Text>}
-        fullWidth
-      >
-        Azure limpide
-      </Button>
-      <Button
-        variant="solid"
-        size="md"
-        spinnerPlacement="end"
-        isLoading
-        themeColor="success"
-        startContent={<Text>🚀</Text>}
-        fullWidth
-      >
-        Azure limpide
-      </Button>
-      <Button
-        variant="solid"
+      <Select
+        placeholder="Select an option"
         size="lg"
-        spinnerPlacement="end"
-        themeColor="warning"
-        startContent={<Text>🚀</Text>}
+        onSelectionChange={keys => console.log(keys)}
         fullWidth
+        selectionMode="multiple"
+        labelPlacement="outside-top"
+        label="Select an option"
+        variant="faded"
       >
-        Azure limpide
-      </Button>
-      <Progress value={0.5} />
-      <Progress value={0.5} variant="circular" />
-      <ActivityIndicator variant="linear" showTrack />
-      <ActivityIndicator showTrack />
-      <Checkbox label="Accept Terms and Conditions" themeColor="primary" />
-      <Checkbox
-        label="Accept Terms and Conditions"
-        themeColor="secondary"
-        variant="light"
-        labelAlignment="justify-right"
-      />
-      <Checkbox
-        fullWidth
-        label="Accept Terms and Conditions"
-        labelAlignment="justify-left"
-        themeColor="success"
-      />
-      <Switch
+        <SelectItem key="option-1" value="option-1" label="Option 1" />
+        <SelectItem key="option-2" value="option-2" label="Option 2" />
+        <SelectItem key="option-3" value="option-3" label="Option 3" />
+      </Select>
+      <Button
+        size="lg"
+        onPress={() => console.log('pressed')}
+        variant="solid"
         themeColor="primary"
-        variant="overlap"
-        label="hello"
-        labelAlignment="justify-left"
-      />
-      <Switch themeColor="secondary" />
-      <Switch themeColor="success" />
-      <Checkbox label="Accept Terms and Conditions" themeColor="warning" />
+      >
+        Press me
+      </Button>
     </View>
   )
 }