+++ packages/native/src/core/theme-hooks.ts
@@ -1,6 +1,7 @@
-import { useContext } from 'react'
+import { useContext, useMemo } from 'react'
 import { useColorScheme } from 'react-native'
 import { XUIThemeContext } from './theme-context'
+import { Radius } from '../types'
 import { XUITheme } from '@xaui/core/theme'
 
 type ColorMode = 'light' | 'dark'
@@ -24,4 +25,26 @@ export function useXUIColors(): XUITheme['colors'] {
   const theme = useXUITheme()
 
   return theme.colors
+}
+
+export function useXUIPalette(): XUITheme['palette'] {
+  const theme = useXUITheme()
+
+  return useMemo(() => theme.palette, [theme])
+}
+
+export function useBorderRadiusStyles(radius: Radius): { borderRadius: number } {
+  const theme = useXUITheme()
+  const borderRadius = useMemo(() => {
+    const radiusMap = {
+      none: theme.borderRadius.none,
+      sm: theme.borderRadius.sm,
+      md: theme.borderRadius.md,
+      lg: theme.borderRadius.lg,
+      full: theme.borderRadius.full,
+    }
+    return { borderRadius: radiusMap[radius] }
+  }, [radius, theme])
+
+  return borderRadius
 }