+++ packages/native/src/index.ts
@@ -1,1 +1,2 @@
 export { ActivityIndicator } from './components/indicator'
+export { Divider } from './components/divider'