+++ packages/native/src/types/index.ts
@@ -1,1 +1,27 @@
-export * from './theme'
+export type ThemeColor =
+  | 'primary'
+  | 'secondary'
+  | 'tertiary'
+  | 'danger'
+  | 'warning'
+  | 'success'
+  | 'default'
+
+export type Size = 'xs' | 'sm' | 'md' | 'lg'
+
+export type Radius = 'none' | 'sm' | 'md' | 'lg' | 'full'
+
+export type SizeValues = {
+  xs: number
+  sm: number
+  md: number
+  lg: number
+}
+
+export type RadiusValues = {
+  none: number
+  sm: number
+  md: number
+  lg: number
+  full: number
+}