+++ packages/hybrid/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @xaui/web
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [d05a489]
+  - @xaui/core@0.1.7
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [c09ae37]
+  - @xaui/core@0.1.6
+
 ## 0.0.2
 
 ### Patch Changes